refactor: modularize 2.8K-line chatgpt-web.ts (Issue #3594)
- extracted session, sentinel, thinking, messages, sse, responses, images, and utils to domain files
- chatgpt-web.ts is now a 1-line facade re-exporting modules
