+++ open-sse/executors/chatgpt-web/session.ts
@@ -1,0 +1,226 @@
+import { BaseExecutor, type ExecuteInput, type ProviderCredentials } from "../base.ts";
+import { OMNIROUTE_VERSION } from "@/shared/constants/version.ts";
+import { getProxyForAccount } from "../../utils/proxyFallback.ts";
+import { HttpsProxyAgent } from "https-proxy-agent";
+import crypto from "node:crypto";
+import { createHash } from "node:crypto";
+import { saveCallLog } from "@/lib/usage/callLogArtifacts.ts";
+import { streamWithTimeout } from "../../utils/stream.ts";
+import { ANTIGRAVITY_CONFIG } from "../../config/errorConfig.ts";
+import { storeChatGptImage, getChatGptImageConversationContext, __resetChatGptImageCacheForTesting, type ChatGptImageConversationContext } from "../../services/chatgptImageCache.ts";
+
+import { SESSION_URL, browserHeaders } from "./constants.ts";
+import { warmupCache } from "./warmup.ts";
+
+// ─── Session token cache ────────────────────────────────────────────────────
+
+export interface TokenEntry {
+  accessToken: string;
+  accountId: string | null;
+  expiresAt: number;
+  refreshedCookie?: string;
+}
+
+export const TOKEN_TTL_MS = 5 * 60 * 1000;
+
+// 5min — accessTokens are short-lived
+export const tokenCache = new Map<string, TokenEntry>();
+
+export function cookieKey(cookie: string): string {
+  // SHA-256 prefix (64 bits). Used as the Map key for tokenCache and
+  // warmupCache; the previous 32-bit FNV-1a was small enough that a
+  // birthday-paradox collision could surface one user's cached accessToken
+  // to another's request. 64 bits is overkill for the 200-entry cache but
+  // costs essentially nothing.
+  // Not a password hash — SHA-256 is used to derive a short, collision-resistant
+  // cache key from the session cookie. The output is a map lookup key.
+  return createHash("sha256").update(cookie).digest("hex").slice(0, 16); // lgtm[js/insufficient-password-hash]
+}
+
+export function tokenLookup(cookie: string): TokenEntry | null {
+  const entry = tokenCache.get(cookieKey(cookie));
+  if (!entry) return null;
+  if (Date.now() >= entry.expiresAt) {
+    tokenCache.delete(cookieKey(cookie));
+    return null;
+  }
+  return entry;
+}
+
+export const TOKEN_CACHE_MAX = 200;
+
+export function tokenStore(cookie: string, entry: TokenEntry): void {
+  // Bound the cache to TOKEN_CACHE_MAX entries (FIFO). Same shape as the
+  // image cache and warmup cache — drop the oldest before inserting.
+  if (tokenCache.size >= TOKEN_CACHE_MAX && !tokenCache.has(cookieKey(cookie))) {
+    const firstKey = tokenCache.keys().next().value;
+    if (firstKey) tokenCache.delete(firstKey);
+  }
+  tokenCache.set(cookieKey(cookie), entry);
+}
+
+// Conversation continuity is intentionally not cached. Open WebUI and most
+// OpenAI-API-style clients re-send the full history each turn, so each
+// request just starts a fresh conversation. Temporary Chat mode is the
+// default; it gets disabled per-request only for image-gen prompts, since
+// that mode rejects the image_gen tool.
+
+// ─── /api/auth/session — exchange cookie for JWT ────────────────────────────
+
+export interface SessionResponse {
+  accessToken?: string;
+  expires?: string;
+  user?: { id?: string };
+}
+
+// Session-token family — NextAuth uses one of these depending on token size:
+//   __Secure-next-auth.session-token            (unchunked, < 4KB)
+//   __Secure-next-auth.session-token.0          (chunked, first piece)
+//   __Secure-next-auth.session-token.N          (chunked, additional pieces)
+// Rotation can change the shape (unchunked → chunked or vice versa). When
+// that happens, every old family member must be dropped — keeping the stale
+// variant alongside the new one would send both, and depending on parser
+// precedence the server could read the stale value and fail auth.
+export const SESSION_TOKEN_FAMILY_RE = /^__Secure-next-auth\.session-token(?:\.\d+)?$/;
+
+/**
+ * Merge any rotated session-token chunks from a Set-Cookie response into the
+ * original cookie blob, preserving every other cookie the caller pasted
+ * (cf_clearance, __cf_bm, _cfuvid, _puid, ...). Returns null if no rotation
+ * occurred or the rotated chunks match what's already there.
+ *
+ * Returning only the matched session-token chunks here was a bug: when the
+ * caller pastes a full DevTools Cookie line (the recommended form), the
+ * Cloudflare cookies are required for subsequent requests, and dropping
+ * them re-triggers `cf-mitigated: challenge`.
+ */
+export function mergeRefreshedCookie(
+  originalCookie: string,
+  setCookieHeader: string | null
+): string | null {
+  if (!setCookieHeader) return null;
+  const matches = Array.from(
+    setCookieHeader.matchAll(/(__Secure-next-auth\.session-token(?:\.\d+)?)=([^;,\s]+)/g)
+  );
+  if (matches.length === 0) return null;
+
+  const refreshed = new Map<string, string>();
+  for (const m of matches) refreshed.set(m[1], m[2]);
+
+  let blob = originalCookie.trim();
+  if (/^cookie\s*:\s*/i.test(blob)) blob = blob.replace(/^cookie\s*:\s*/i, "");
+
+  // Bare value (no `=`): the original was just the session-token contents.
+  // Replace with the new chunked form.
+  if (!/=/.test(blob)) {
+    return Array.from(refreshed, ([k, v]) => `${k}=${v}`).join("; ");
+  }
+
+  const pairs = blob.split(/;\s*/).filter(Boolean);
+  const result: string[] = [];
+  let mutated = false;
+  let droppedStale = false;
+  for (const pair of pairs) {
+    const eqIdx = pair.indexOf("=");
+    if (eqIdx < 0) {
+      result.push(pair);
+      continue;
+    }
+    const name = pair.slice(0, eqIdx).trim();
+    const value = pair.slice(eqIdx + 1);
+    // Drop ALL session-token-family members from the original — we'll
+    // append the refreshed set below. This handles unchunked→chunked and
+    // chunked→unchunked rotations, where keeping the old name would leave
+    // the stale token visible alongside the new one.
+    if (SESSION_TOKEN_FAMILY_RE.test(name)) {
+      if (!refreshed.has(name) || refreshed.get(name) !== value) mutated = true;
+      droppedStale = true;
+      continue;
+    }
+    result.push(`${name}=${value}`);
+  }
+  // Append the full refreshed family.
+  for (const [name, value] of refreshed) {
+    result.push(`${name}=${value}`);
+  }
+  if (!droppedStale) mutated = true; // refreshed chunks were entirely new
+  return mutated ? result.join("; ") : null;
+}
+
+/**
+ * Build the Cookie header value from whatever the user pasted.
+ *
+ * Accepts:
+ *   - A bare value:                       "eyJhbGc..."  →  prepended with __Secure-next-auth.session-token=
+ *   - An unchunked cookie line:           "__Secure-next-auth.session-token=eyJ..."
+ *   - A chunked cookie line:              "__Secure-next-auth.session-token.0=...; __Secure-next-auth.session-token.1=..."
+ *   - The full DevTools cookie header:    "Cookie: __Secure-next-auth.session-token.0=...; cf_clearance=..."
+ *
+ * If the user pastes a chunked token, we pass the cookies through verbatim —
+ * NextAuth's server reassembles them on its side.
+ */
+export function buildSessionCookieHeader(rawInput: string): string {
+  let s = rawInput.trim();
+  if (/^cookie\s*:\s*/i.test(s)) s = s.replace(/^cookie\s*:\s*/i, "");
+  if (/__Secure-next-auth\.session-token(?:\.\d+)?\s*=/.test(s)) {
+    return s;
+  }
+  return `__Secure-next-auth.session-token=${s}`;
+}
+
+async export function exchangeSession(
+  cookie: string,
+  signal: AbortSignal | null | undefined
+): Promise<TokenEntry> {
+  const cached = tokenLookup(cookie);
+  if (cached) return cached;
+
+  const headers: Record<string, string> = {
+    ...browserHeaders(),
+    Accept: "application/json",
+    Cookie: buildSessionCookieHeader(cookie),
+  };
+
+  const response = await tlsFetchChatGpt(SESSION_URL, {
+    method: "GET",
+    headers,
+    timeoutMs: 30_000,
+    signal,
+  });
+
+  if (response.status === 401 || response.status === 403) {
+    throw new SessionAuthError("Invalid session cookie");
+  }
+  if (response.status >= 400) {
+    throw new Error(`Session exchange failed (HTTP ${response.status})`);
+  }
+
+  const refreshed = mergeRefreshedCookie(cookie, response.headers.get("set-cookie"));
+  let data: SessionResponse = {};
+  try {
+    data = JSON.parse(response.text || "{}");
+  } catch {
+    console.warn("[chatgpt-web] session response JSON parse failed");
+    /* empty body or non-JSON */
+  }
+  if (!data.accessToken) {
+    throw new SessionAuthError("Session response missing accessToken — cookie likely expired");
+  }
+
+  const expiresAt = data.expires ? new Date(data.expires).getTime() : Date.now() + TOKEN_TTL_MS;
+  const entry: TokenEntry = {
+    accessToken: data.accessToken,
+    accountId: data.user?.id ?? null,
+    expiresAt: Math.min(expiresAt, Date.now() + TOKEN_TTL_MS),
+    refreshedCookie: refreshed ?? undefined,
+  };
+  tokenStore(cookie, entry);
+  return entry;
+}
+
+export class SessionAuthError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "SessionAuthError";
+  }
+}
