+++ open-sse/executors/chatgpt-web/warmup.ts
@@ -1,0 +1,79 @@
+import { BaseExecutor, type ExecuteInput, type ProviderCredentials } from "../base.ts";
+import { OMNIROUTE_VERSION } from "@/shared/constants/version.ts";
+import { getProxyForAccount } from "../../utils/proxyFallback.ts";
+import { HttpsProxyAgent } from "https-proxy-agent";
+import crypto from "node:crypto";
+import { createHash } from "node:crypto";
+import { saveCallLog } from "@/lib/usage/callLogArtifacts.ts";
+import { streamWithTimeout } from "../../utils/stream.ts";
+import { ANTIGRAVITY_CONFIG } from "../../config/errorConfig.ts";
+import { storeChatGptImage, getChatGptImageConversationContext, __resetChatGptImageCacheForTesting, type ChatGptImageConversationContext } from "../../services/chatgptImageCache.ts";
+
+import { CHATGPT_BASE, browserHeaders, oaiHeaders } from "./constants.ts";
+import { cookieKey, buildSessionCookieHeader } from "./session.ts";
+
+// ─── Session warmup ────────────────────────────────────────────────────────
+// Mimics chatgpt.com's page-load fetch sequence so Sentinel sees a "warm"
+// browsing session. Cached per (cookie, access-token) pair for 60s to avoid
+// hammering the warmup endpoints on every chat completion.
+
+export const warmupCache = new Map<string, number>();
+
+export const WARMUP_TTL_MS = 60_000;
+
+export const WARMUP_CACHE_MAX = 200;
+
+async export function runSessionWarmup(
+  accessToken: string,
+  accountId: string | null,
+  sessionId: string,
+  deviceId: string,
+  cookie: string,
+  signal: AbortSignal | null | undefined,
+  log: { debug?: (tag: string, msg: string) => void } | null | undefined
+): Promise<void> {
+  const key = cookieKey(cookie) + ":" + accessToken.slice(-8);
+  const now = Date.now();
+  const last = warmupCache.get(key);
+  if (last && now - last < WARMUP_TTL_MS) return;
+  // Bound the cache: drop the oldest entry once we hit the cap. Map iteration
+  // order is insertion order, so the first key is the oldest.
+  if (warmupCache.size >= WARMUP_CACHE_MAX && !warmupCache.has(key)) {
+    const first = warmupCache.keys().next().value;
+    if (first) warmupCache.delete(first);
+  }
+  warmupCache.set(key, now);
+
+  const headers: Record<string, string> = {
+    ...browserHeaders(),
+    ...oaiHeaders(sessionId, deviceId),
+    Accept: "*/*",
+    Authorization: `Bearer ${accessToken}`,
+    Cookie: buildSessionCookieHeader(cookie),
+    Priority: "u=1, i",
+  };
+  if (accountId) headers["chatgpt-account-id"] = accountId;
+
+  const urls = [
+    `${CHATGPT_BASE}/backend-api/me`,
+    `${CHATGPT_BASE}/backend-api/conversations?offset=0&limit=28&order=updated`,
+    `${CHATGPT_BASE}/backend-api/models?history_and_training_disabled=false`,
+  ];
+
+  for (const url of urls) {
+    try {
+      const r = await tlsFetchChatGpt(url, {
+        method: "GET",
+        headers,
+        timeoutMs: 15_000,
+        signal,
+      });
+      log?.debug?.("CGPT-WEB", `warmup ${url.split("/backend-api/")[1]} → ${r.status}`);
+    } catch (err) {
+      log?.debug?.(
+        "CGPT-WEB",
+        `warmup ${url} failed: ${err instanceof Error ? err.message : String(err)}`
+      );
+    }
+  }
+}
